pyproject.toml
[tool.poetry]
name = "overtime-machine"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "overtime_machine"}]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "<2.1.1"
openpyxl = "^3.1.2"
python-docx = "^1.0.1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
